requirements.txt
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.3.2
onnxruntime==1.22.1
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
PyAudio==0.2.14
pyodbc==5.2.0
pypyodbc==1.3.6
pyreadline3==3.5.4
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.22.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
